requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
certifi==2022.6.15; python_version >= '3.6'
django==4.1
djangorestframework==3.13.1
gunicorn==20.1.0
pillow==9.2.0
psycopg2==2.9.3
pytz==2022.2.1
sentry-sdk==1.9.7
setuptools==65.3.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
urllib3==1.26.12; python_version >= '3.6'
